fix: resolve circular imports and move constants to db.constants
