+++ rotkehlchen/db/repositories/session_management.py
@@ -3,7 +3,7 @@ import logging
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from rotkehlchen.db.dbhandler import DBINFO_FILENAME
+from rotkehlchen.db.constants import DBINFO_FILENAME
 from rotkehlchen.errors.misc import SystemPermissionError
 from rotkehlchen.utils.hashing import file_md5
 from rotkehlchen.utils.serialization import rlk_jsondumps
